backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-environ>=0.12.0",
    "django-jsonform>=2.23.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg>=3.3.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]


[tool.ruff]
line-length = 100
extend-exclude = [
    "**/migrations/**"
]

[tool.ruff.lint]
select = [
    "I" # Isort
]
